Move lowering to mathimpl and address formatting issues
